@@ -6,15 +6,15 @@ import { BaseService } from "@/api/base.service";
  * API abstraction layer communication via Axios (typescript singleton pattern)
  */
 class EmployeeService extends BaseService {
+  private static controller = "Employee";
   private static employeeService: EmployeeService;
-  private static apiBaseUrl = "http://localhost:55276/api/Employee/";
 
   private constructor(apiBaseUrl: string) {
     super(apiBaseUrl);
   }
 
   public static get Instance(): EmployeeService {
-    return (this.employeeService || (this.employeeService = new this(this.apiBaseUrl)));
+    return (this.employeeService || (this.employeeService = new this(this.controller)));
   }
 
   public async getAllEmployeesAsync(): Promise<IEmployee[]> {
@@ -24,21 +24,21 @@ class EmployeeService extends BaseService {
 
   public async addEmployeeAsync(employee: IEmployee): Promise<any> {
     const config: AxiosRequestConfig = {
-      params: { id: employee.Id }
+      params: { id: employee.id }
     };
     await this.$http.post("AddEmployeeAsync", employee, config);
   }
 
   public async deleteEmployeeAsync(employee: IEmployee): Promise<any> {
     const config: AxiosRequestConfig = {
-      params: { id: employee.Id }
+      params: { id: employee.id }
     };
     await this.$http.delete("DeleteEmployeeAsync", config);
   }
 
   public async getEmployeeByIdAsync(id: number | null = null): Promise<IEmployee> {
     const config: AxiosRequestConfig = {
-      params: { id: id }
+      params: { id }
     };
     const { data } = await this.$http.get<IEmployee>("GetEmployeeByIdAsync", config);
     return data;
