@@ -21,18 +21,17 @@ namespace FullStackTesting.Web.Api.IntegrationTests.Controllers
         public async Task CanGetAllEmployeesAsync()
         {
             // The endpoint or route of the controller action
-            using (var httpResponse = await _client.GetAsync("/api/Employee/GetAllEmployeesAsync"))
-            {
-                // Must be successful
-                httpResponse.EnsureSuccessStatusCode();
+            var httpResponse = await _client.GetAsync("/api/Employee/GetAllEmployeesAsync");
 
-                // Deserialize and examine results
-                var stringResponse = await httpResponse.Content.ReadAsStringAsync();
-                var employees = JsonConvert.DeserializeObject<List<Employee>>(stringResponse);
+            // Must be successful
+            httpResponse.EnsureSuccessStatusCode();
 
-                Assert.Contains(employees, e => e.FirstName.Equals("Matt") && e.LastName.Equals("Areddia"));
-                Assert.Contains(employees, e => e.FirstName.Equals("Jeremy") && e.LastName.Equals("Wu"));
-            }
+            // Deserialize and examine results
+            var stringResponse = await httpResponse.Content.ReadAsStringAsync();
+            var employees = JsonConvert.DeserializeObject<List<Employee>>(stringResponse);
+
+            Assert.Contains(employees, e => e.FirstName.Equals("Matt") && e.LastName.Equals("Areddia"));
+            Assert.Contains(employees, e => e.FirstName.Equals("Jeremy") && e.LastName.Equals("Wu"));
         }
 
         [Fact]
@@ -40,19 +39,18 @@ namespace FullStackTesting.Web.Api.IntegrationTests.Controllers
         {
             // The endpoint or route of the controller action
             const int targetId = 5;
-            using (var httpResponse = await _client.GetAsync($"/api/Employee/GetEmployeeByIdAsync?id={targetId}"))
-            {
-                // Must be successful
-                httpResponse.EnsureSuccessStatusCode();
+            var httpResponse = await _client.GetAsync($"/api/Employee/GetEmployeeByIdAsync?id={targetId}");
 
-                // Deserialize and examine results
-                var stringResponse = await httpResponse.Content.ReadAsStringAsync();
-                var employee = JsonConvert.DeserializeObject<Employee>(stringResponse);
+            // Must be successful
+            httpResponse.EnsureSuccessStatusCode();
 
-                Assert.Equal(targetId, employee.Id);
-                Assert.Equal("Jeremy", employee.FirstName);
-                Assert.Equal("Wu", employee.LastName);
-            }
+            // Deserialize and examine results
+            var stringResponse = await httpResponse.Content.ReadAsStringAsync();
+            var employee = JsonConvert.DeserializeObject<Employee>(stringResponse);
+
+            Assert.Equal(targetId, employee.Id);
+            Assert.Equal("Jeremy", employee.FirstName);
+            Assert.Equal("Wu", employee.LastName);
         }
 
         [Fact]
@@ -70,19 +68,18 @@ namespace FullStackTesting.Web.Api.IntegrationTests.Controllers
 
             // The endpoint or route of the controller action (AddEmployeeAsync) with StringContent comprised of the employee to add
             var addEmployeeStringContent = new StringContent(JsonConvert.SerializeObject(addEmployee), Encoding.UTF8, "application/json");
-            using (var httpResponse = await _client.PostAsync($"/api/Employee/AddEmployeeAsync?id={addEmployee.Id}", addEmployeeStringContent))
-            {
-                // Must be successful
-                httpResponse.EnsureSuccessStatusCode();
+            var httpResponse = await _client.PostAsync($"/api/Employee/AddEmployeeAsync?id={addEmployee.Id}", addEmployeeStringContent);
 
-                // Deserialize and examine results (compare employee object returned in response to that passed in initial request - should match)
-                var stringResponse = await httpResponse.Content.ReadAsStringAsync();
-                var employee = JsonConvert.DeserializeObject<Employee>(stringResponse);
+            // Must be successful
+            httpResponse.EnsureSuccessStatusCode();
 
-                Assert.Equal(addEmployee.Id, employee.Id);
-                Assert.Equal(addEmployee.FirstName, employee.FirstName);
-                Assert.Equal(addEmployee.LastName, employee.LastName);
-            }
+            // Deserialize and examine results (compare employee object returned in response to that passed in initial request - should match)
+            var stringResponse = await httpResponse.Content.ReadAsStringAsync();
+            var employee = JsonConvert.DeserializeObject<Employee>(stringResponse);
+
+            Assert.Equal(addEmployee.Id, employee.Id);
+            Assert.Equal(addEmployee.FirstName, employee.FirstName);
+            Assert.Equal(addEmployee.LastName, employee.LastName);
         }
 
         [Fact]
@@ -100,27 +97,25 @@ namespace FullStackTesting.Web.Api.IntegrationTests.Controllers
 
             // The endpoint or route of the controller action (UpdateEmployeeAsync) with StringContent comprised of the employee to update / id of employee to update
             var updateEmployeeStringContent = new StringContent(JsonConvert.SerializeObject(updateEmployee), Encoding.UTF8, "application/json");
+
+            // Must be successful
             using (var httpPutResponse = await _client.PutAsync($"/api/Employee/UpdateEmployeeAsync?id={updateEmployee.Id}", updateEmployeeStringContent))
-            {
-                // Must be successful
                 httpPutResponse.EnsureSuccessStatusCode();
-            }
 
             // The endpoint or route of the controller action (GetEmployeeByIdAsync) - the employee we updated
-            using (var httpResponse = await _client.GetAsync($"/api/Employee/GetEmployeeByIdAsync?id={updateEmployee.Id}"))
-            {
-                // Must be successful
-                httpResponse.EnsureSuccessStatusCode();
+            var httpResponse = await _client.GetAsync($"/api/Employee/GetEmployeeByIdAsync?id={updateEmployee.Id}");
 
-                // Deserialize and examine results
-                var stringResponse = await httpResponse.Content.ReadAsStringAsync();
-                var employee = JsonConvert.DeserializeObject<Employee>(stringResponse);
+            // Must be successful
+            httpResponse.EnsureSuccessStatusCode();
 
-                Assert.Equal(updateEmployee.FirstName, employee.FirstName);
-                Assert.Equal(updateEmployee.LastName, employee.LastName);
-                Assert.Equal(updateEmployee.FullTime, employee.FullTime);
-                Assert.Equal(updateEmployee.Department, employee.Department);
-            }
+            // Deserialize and examine results
+            var stringResponse = await httpResponse.Content.ReadAsStringAsync();
+            var employee = JsonConvert.DeserializeObject<Employee>(stringResponse);
+
+            Assert.Equal(updateEmployee.FirstName, employee.FirstName);
+            Assert.Equal(updateEmployee.LastName, employee.LastName);
+            Assert.Equal(updateEmployee.FullTime, employee.FullTime);
+            Assert.Equal(updateEmployee.Department, employee.Department);
         }
 
         [Fact]
@@ -128,24 +123,22 @@ namespace FullStackTesting.Web.Api.IntegrationTests.Controllers
         {
             // The endpoint or route of the controller action (DeleteEmployeeAsync)
             const int targetId = 2;
+
+            // Must be successful
             using (var httpDeleteResponse = await _client.DeleteAsync($"/api/Employee/DeleteEmployeeAsync?id={targetId}"))
-            {
-                // Must be successful
                 httpDeleteResponse.EnsureSuccessStatusCode();
-            }
 
             // The endpoint or route of the controller action (GetAllEmployeesAsync)
-            using (var httpResponse = await _client.GetAsync("/api/Employee/GetAllEmployeesAsync"))
-            {
-                // Must be successful
-                httpResponse.EnsureSuccessStatusCode();
+            var httpResponse = await _client.GetAsync("/api/Employee/GetAllEmployeesAsync");
 
-                // Deserialize and examine results (should no longer contain record having Id = 1, FirstName = 'Matt', LastName = 'Areddia')
-                var stringResponse = await httpResponse.Content.ReadAsStringAsync();
-                var employees = JsonConvert.DeserializeObject<List<Employee>>(stringResponse);
+            // Must be successful
+            httpResponse.EnsureSuccessStatusCode();
 
-                Assert.DoesNotContain(employees, e => e.Id.Equals(targetId) && e.FirstName.Equals("Jane") && e.LastName.Equals("Doe"));
-            }
+            // Deserialize and examine results (should no longer contain record having Id = 1, FirstName = 'Matt', LastName = 'Areddia')
+            var stringResponse = await httpResponse.Content.ReadAsStringAsync();
+            var employees = JsonConvert.DeserializeObject<List<Employee>>(stringResponse);
+
+            Assert.DoesNotContain(employees, e => e.Id.Equals(targetId) && e.FirstName.Equals("Jane") && e.LastName.Equals("Doe"));
         }
     }
 }
